model selection list, more in descrpition
So this is basically, just for a template for later porpuses.
Template was generated with gpt, so it still need some correction according to the project files, some expressions probably not works properly, like: 'user_maganer.current_username'
models.py use a basic model settings in its 'generate_response' algorithm. - Probably it will be not needed at all.